fix migration part

+++ docker-stacks/02-keycloak/etc/hooks/after-start.d/10-configure-simva-realm.sh
@@ -22,41 +22,43 @@ if [[ ${SIMVA_KEYCLOAK_VERSION%%.*} > 18 ]]; then
         fi;
     fi;
     
-    source "${STACK_HOME}/etc/hooks/helpers.d/keycloak-functions.sh"
-    source "${SIMVA_HOME}/bin/get-or-generate.sh"
+    if [[ ! -e "${SIMVA_CONFIG_HOME}/keycloak/simva-realm-export/.exportinprogress" ]]; then 
+        source "${STACK_HOME}/etc/hooks/helpers.d/keycloak-functions.sh"
+        source "${SIMVA_HOME}/bin/get-or-generate.sh"
 
-    __update_realm_with_params -s registrationAllowed=${SIMVA_SSO_SELF_REGISTRATION_ALLOWED}
-    
-    __add_or_update_role "${SIMVA_CONFIG_HOME}/keycloak/simva-realm/roles" "/opt/keycloak/data/simva-realm-filled/roles"
-    __add_or_update_user "${SIMVA_CONFIG_HOME}/keycloak/simva-realm/users" "/opt/keycloak/data/simva-realm-filled/users"
+        __update_realm_with_params -s registrationAllowed=${SIMVA_SSO_SELF_REGISTRATION_ALLOWED}
+        
+        __add_or_update_role "${SIMVA_CONFIG_HOME}/keycloak/simva-realm/roles" "/opt/keycloak/data/simva-realm-filled/roles"
+        __add_or_update_user "${SIMVA_CONFIG_HOME}/keycloak/simva-realm/users" "/opt/keycloak/data/simva-realm-filled/users"
 
-    __keycloak_login
-    # Update users config
-    users="student teaching_assistant teacher researcher administrator"
-    conf_file="${STACK_CONF}/realm-data.users.yml"
-    if [[ -e ${conf_file} ]]; then
-        rm -rf ${conf_file}
-    fi
-    echo "users:" > ${conf_file}
-    for user in $users; do
-        user=$(echo ${user} | sed -e 's/[^0-9A-Za-z_]/_/g' )
-        user_username=$(get_or_generate_username "${user}" "${STACK_CONF}/simva-env.sh" "USER")
-        user_password=$(get_or_generate_password "${user}" "${STACK_CONF}/simva-env.sh" "USER")
-        cat << EOF >> ${conf_file}
+        __keycloak_login
+        # Update users config
+        users="student teaching_assistant teacher researcher administrator"
+        conf_file="${STACK_CONF}/realm-data.users.yml"
+        if [[ -e ${conf_file} ]]; then
+            rm -rf ${conf_file}
+        fi
+        echo "users:" > ${conf_file}
+        for user in $users; do
+            user=$(echo ${user} | sed -e 's/[^0-9A-Za-z_]/_/g' )
+            user_username=$(get_or_generate_username "${user}" "${STACK_CONF}/simva-env.sh" "USER")
+            user_password=$(get_or_generate_password "${user}" "${STACK_CONF}/simva-env.sh" "USER")
+            cat << EOF >> ${conf_file}
   ${user}:
     username: "${user_username}"
     password: "${user_password}"
 EOF
-        echo "Setting password for username ${user_username}"
-        "${SIMVA_HOME}/bin/run-command.sh" /opt/keycloak/bin/kcadm.sh set-password -r ${SIMVA_SSO_REALM} --username $user_username --new-password $user_password
-        echo "Setting password for username ${user_username} done"
-    done
-    
-    __add_or_update_client_scope "${SIMVA_CONFIG_TEMPLATE_HOME}/keycloak/simva-realm/clients-scopes/lti" "/opt/keycloak/data/simva-realm/clients-scopes/lti"
-    __add_or_update_client_scope "${SIMVA_CONFIG_TEMPLATE_HOME}/keycloak/simva-realm/clients-scopes/saml" "/opt/keycloak/data/simva-realm/clients-scopes/saml"
-    __add_or_update_client_scope "${SIMVA_CONFIG_TEMPLATE_HOME}/keycloak/simva-realm/clients-scopes/policy-role" "/opt/keycloak/data/simva-realm/clients-scopes/policy-role"
-    
-    __add_or_update_client "${SIMVA_CONFIG_HOME}/keycloak/simva-realm/clients" "/opt/keycloak/data/simva-realm-filled/clients"
+            echo "Setting password for username ${user_username}"
+            "${SIMVA_HOME}/bin/run-command.sh" /opt/keycloak/bin/kcadm.sh set-password -r ${SIMVA_SSO_REALM} --username $user_username --new-password $user_password
+            echo "Setting password for username ${user_username} done"
+        done
+        
+        __add_or_update_client_scope "${SIMVA_CONFIG_TEMPLATE_HOME}/keycloak/simva-realm/clients-scopes/lti" "/opt/keycloak/data/simva-realm/clients-scopes/lti"
+        __add_or_update_client_scope "${SIMVA_CONFIG_TEMPLATE_HOME}/keycloak/simva-realm/clients-scopes/saml" "/opt/keycloak/data/simva-realm/clients-scopes/saml"
+        __add_or_update_client_scope "${SIMVA_CONFIG_TEMPLATE_HOME}/keycloak/simva-realm/clients-scopes/policy-role" "/opt/keycloak/data/simva-realm/clients-scopes/policy-role"
+        
+        __add_or_update_client "${SIMVA_CONFIG_HOME}/keycloak/simva-realm/clients" "/opt/keycloak/data/simva-realm-filled/clients"
+    fi
 else
     echo "Please upgrade to a newer keycloak version ( > 18.*.*) before importing users and realm."
 fi;
